requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
wordcloud==1.9.3
pillow==10.4.0
python-dotenv==1.2.1

kiwipiepy